backend/requirements.txt
asgiref==3.11.0
dj-database-url==2.3.0
Django==5.1.15
django-cors-headers==4.9.0
djangorestframework==3.15.2
gunicorn==22.0.0
packaging==25.0
psycopg==3.3.2
psycopg-binary==3.3.2
python-dotenv==1.2.1
sqlparse==0.5.5
typing_extensions==4.15.0
whitenoise==6.11.0

